Models/requirements.txt
Flask
requests
yt-dlp
opencv-python
torch
torchvision
scikit-learn
Pillow
